Separate memmove tests from memcpy tests
Move all memmove tests into a new src/test/func/memmove/ directory
so they build as func-memmove-fast and func-memmove-check, separate
from func-memcpy-fast and func-memcpy-check. This makes it clear
which test is failing when CI reports errors.
